set perPage with useScreenWidth hook
